change to general types
