--- cuenca_validations/authorizer/__init__.py
@@ -1,3 +1,0 @@
-__all__ = ['EcommerceIndicator']
-
-from .enums import EcommerceIndicator
