+++ cuenca_validations/types/enums.py
@@ -192,3 +192,15 @@ class SavingCategory(str, Enum):
     medical = 'medical'
     accident = 'accident'
     education = 'education'
+
+
+class EcommerceIndicator(str, Enum):
+    """
+    Adquirente graba
+    Emisor lee
+    0, “ “ = No es una transacción de comercio electrónico
+    1 = Transacción MOTO
+    5 = Comercio seguro, titular autenticado (3D Secure)
+    6 = Comercio seguro, titular no autenticado (3D Secure)
+    7 = Autenticación 3D Secure no realizada
+    """
